refactor(store): minor code-review cleanups
- Rename StoreOrderTests to StoreEntityDefaultsTests (covers
  StoreOrder, StoreProduct and StoreTreasurySyncState defaults).
- Extract OrderCounterpartyInput record into its own file under
  Application/Services/Store/Dtos.
- Drop unused 'year' parameter on IStoreRepository.GetAllOrders;
  rename to GetAllOrdersAsync. Year filtering will be reintroduced
  when the year-handle resolution lands in Phase 5.
- Convert StoreTreasurySyncState.SyncStatus from string to enum
  StoreTreasurySyncStatus { Idle, Running, Failed }, matching the
  HasConversion<int>() pattern used by StoreOrderState and
  StorePaymentMethod. AddStoreSection migration was regenerated
  to pick up the int column type (the prior, unmerged migration
  was removed cleanly via dotnet ef migrations remove).
