+++ src/Humans.Application/Services/Store/Dtos/OrderCounterpartyInput.cs
@@ -1,0 +1,8 @@
+namespace Humans.Application.Services.Store.Dtos;
+
+public record OrderCounterpartyInput(
+    string? Name,
+    string? VatId,
+    string? Address,
+    string? CountryCode,
+    string? Email);
