+++ src/Humans.Domain/Entities/StoreTreasurySyncState.cs
@@ -1,3 +1,4 @@
+using Humans.Domain.Enums;
 using NodaTime;
 
 namespace Humans.Domain.Entities;
@@ -6,6 +7,6 @@ public class StoreTreasurySyncState
 {
     public int Id { get; set; } = 1;
     public Instant? LastSyncAt { get; set; }
-    public string SyncStatus { get; set; } = "Idle";
+    public StoreTreasurySyncStatus SyncStatus { get; set; } = StoreTreasurySyncStatus.Idle;
     public string? LastError { get; set; }
 }
