+++ src/Humans.Domain/Enums/StoreTreasurySyncStatus.cs
@@ -1,0 +1,8 @@
+namespace Humans.Domain.Enums;
+
+public enum StoreTreasurySyncStatus
+{
+    Idle = 0,
+    Running = 1,
+    Failed = 2
+}
